[modify] requestHandlerを分離
